fix: рефакторинг кода
